+++ resource/commands/restart.sh
@@ -1,9 +1,25 @@
 
 app_name=$(get_option '.app_name')
-staging_timeout=$(get_option '.staging_timeout')
-startup_timeout=$(get_option '.startup_timeout')
+strategy=$(get_option '.strategy')
+no_wait=$(get_option '.no_wait')
+staging_timeout=$(get_option '.staging_timeout' 0)
+startup_timeout=$(get_option '.startup_timeout' 0)
 
 logger::info "Executing $(logger::highlight "$command"): $app_name"
 
 cf::target "$org" "$space"
-cf::restart "$app_name" "$staging_timeout" "$startup_timeout"
+
+[ "$staging_timeout" -gt "0" ] && export CF_STAGING_TIMEOUT=$staging_timeout
+[ "$startup_timeout" -gt "0" ] && export CF_STARTUP_TIMEOUT=$startup_timeout
+
+args=("$app_name")
+
+if ! cf::is_cf6; then
+  [ -n "$strategy" ] && args+=(--strategy "$strategy")
+  [ "$no_wait" == "true"  ] && args+=(--no-wait)
+fi
+
+cf::cf restart "${args[@]}"
+
+unset CF_STAGING_TIMEOUT
+unset CF_STARTUP_TIMEOUT